Moved function to composable
