+++ Idle_Game/app/src/main/java/com/example/idle_game/MainActivity.kt
@@ -21,6 +21,7 @@ import androidx.navigation.compose.rememberNavController
 import com.example.idle_game.ui.navigation.BottomBar
 import com.example.idle_game.ui.navigation.NavigationGraph
 import com.example.idle_game.ui.theme.Idle_GameTheme
+import com.example.idle_game.ui.views.composable.Idle_GameLauncher
 import com.example.idle_game.ui.views.composable.LoginView
 import com.example.idle_game.ui.views.composable.StartView
 import com.example.idle_game.ui.views.models.LoginViewModel
@@ -34,44 +35,9 @@ class MainActivity : ComponentActivity() {
 
         setContent {
            Idle_GameLauncher()
-            //LOGIN aufruf
-            //LoginView(LoginViewModel())
 
-            /* Idle_GameTheme {
-                val navController = rememberNavController()
-                Scaffold (
-                    modifier = Modifier.fillMaxSize(),
-                    bottomBar = { BottomBar(navController = navController) },
-                    content = { it ->
-                        Row(modifier = Modifier.padding(it)) {
-                            NavigationGraph(navController = navController)
-                        }
-                    }
-                )
-            }*/
         }
     }
 }
 
-@Composable
-fun Idle_GameLauncher(modifier: Modifier = Modifier) {
-    var isLoggedIn by remember { mutableStateOf(false) }
 
-    if (isLoggedIn) {
-        Idle_GameTheme {
-            val navController = rememberNavController()
-            Scaffold (
-                modifier = Modifier.fillMaxSize(),
-                bottomBar = { BottomBar(navController = navController) },
-                content = { it ->
-                    Row(modifier = Modifier.padding(it)) {
-                        NavigationGraph(navController = navController)
-                    }
-                }
-            )
-        }
-    } else {
-//        LoginView(LoginViewModel(/*hilt + */onLoginSuccess = { isLoggedIn = true }), onLoginSuccess = { isLoggedIn = true })
-        LoginView(hiltViewModel(), onLoginSuccess = { isLoggedIn = true })//, onLoginSuccess = { isLoggedIn = true })
-    }
-}
